budgetin/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2
django-cors-headers==3.11.0
django-soft-delete==0.9.18
djangorestframework==3.13.1
idna==3.3
psycopg2==2.9.3
pycparser==2.21
pyDes==2.0.1
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
